Move `arrayUtils` into `web-common/lib`
