+++ web-common/src/features/models/utils/embedded.ts
@@ -1,10 +1,10 @@
 import type { Reference } from "@rilldata/web-common/features/models/utils/get-table-references";
 import { humanReadableErrorMessage } from "@rilldata/web-common/features/sources/add-source/errors";
+import { getMapFromArray } from "@rilldata/web-common/lib/arrayUtils";
 import type {
   V1CatalogEntry,
   V1ReconcileError,
 } from "@rilldata/web-common/runtime-client";
-import { getMapFromArray } from "@rilldata/web-local/lib/util/arrayUtils";
 
 export function filterKnownEmbeddedSources(
   embeddedRefs: Array<Reference>,
